+++ Moteur_physique/Vecteur3D.h
@@ -27,7 +27,7 @@ public:
 
 	//Operations
 	Vecteur3D operator+(const Vecteur3D& b);
-	Vecteur3D operator*(const double& lambda);
+	Vecteur3D operator*(const float& lambda);
 
 	Vecteur3D multScalaire(float k);
 	Vecteur3D prodComposantes (Vecteur3D v);
